+++ src/main/java/com/kh/iMMUTABLE/entity/Faq.java
@@ -14,7 +14,7 @@ public class Faq {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "faq_id")
-    private int faqId;
+    private long faqId;
 
     private String faqTitle;
     @Lob
