+++ src/main/java/com/kh/iMMUTABLE/repository/ProductRepository.java
@@ -1,6 +1,7 @@
 package com.kh.iMMUTABLE.repository;
 
 import com.kh.iMMUTABLE.constant.ProductSellStatus;
+import com.kh.iMMUTABLE.dto.ProductDto;
 import com.kh.iMMUTABLE.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -9,7 +10,8 @@ import javax.persistence.PersistenceContext;
 import java.util.List;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    Product findByProductId(int ProductId);
+    Product findByProductId(long ProductId);
+
 
     List<Product> findByProductSellStatus(ProductSellStatus sellStatus);
