+++ src/main/java/com/kh/iMMUTABLE/repository/QnaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface QnaRepository extends JpaRepository<Qna, Long> {
-    Qna findByQnaId(int qnaId);
+    Qna findByQnaId(long qnaId);
 
 }
