+++ src/main/java/com/kh/iMMUTABLE/service/QnaService.java
@@ -28,7 +28,7 @@ public class QnaService {
     }
 
     //save를 통한 update문 기능 findByQnaId를 통해서 select 한 후 set을 통해 update를 자동으로 변환 시켜준다.
-    public boolean upLoadReply(int qnaId,String qnaStatue,String qnaReplay) {
+    public boolean upLoadReply(long qnaId,String qnaStatue,String qnaReplay) {
         System.out.println("서비스 : " +  qnaReplay);
         Qna qna = qnaRepository.findByQnaId(qnaId);
         qna.setQnaStatus(QnaStatus.valueOf(qnaStatue));
